pyproject.toml
[tool.poetry]
name = "flask-template"
version = "0.1.0"
description = "Template for flask projects"
authors = ["Francisco <vieirafrancisco.dev@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Flask = "^3.0.3"
pytest = "^8.3.3"
environs = "^11.0.0"
pypdf2 = "^3.0.1"
google-generativeai = "^0.8.1"
nltk = "^3.9.1"
flask-sqlalchemy = "^3.1.1"
psycopg2-binary = "^2.9.10"

[tool.poetry.group.dev.dependencies]
ruff = "^0.6.5"
flask-shell-ipython = "^0.5.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
